+++ e2e/runner/runner.go
@@ -116,6 +116,16 @@ type E2ERunner struct {
 	GatewayProgram solana.PublicKey
 	SPLAddr        solana.PublicKey
 
+	// contract Sui
+	GatewayPackageID string
+	GatewayObjectID  string
+
+	// SuiTokenCoinType is the coin type identifying the fungible token for SUI
+	SuiTokenCoinType string
+
+	// SuiTokenTreasuryCap is the treasury cap for the SUI token that allows minting, only using in local tests
+	SuiTokenTreasuryCap string
+
 	// contracts evm
 	ZetaEthAddr       ethcommon.Address
 	ZetaEth           *zetaeth.ZetaEth
@@ -132,18 +142,25 @@ type E2ERunner struct {
 	TestDAppV2EVM     *testdappv2.TestDAppV2
 
 	// contracts zevm
-	ERC20ZRC20Addr       ethcommon.Address
-	ERC20ZRC20           *zrc20.ZRC20
-	SPLZRC20Addr         ethcommon.Address
-	SPLZRC20             *zrc20.ZRC20
-	ETHZRC20Addr         ethcommon.Address
-	ETHZRC20             *zrc20.ZRC20
-	BTCZRC20Addr         ethcommon.Address
-	BTCZRC20             *zrc20.ZRC20
-	SOLZRC20Addr         ethcommon.Address
-	SOLZRC20             *zrc20.ZRC20
-	TONZRC20Addr         ethcommon.Address
-	TONZRC20             *zrc20.ZRC20
+	// zrc20 contracts
+	ERC20ZRC20Addr    ethcommon.Address
+	ERC20ZRC20        *zrc20.ZRC20
+	SPLZRC20Addr      ethcommon.Address
+	SPLZRC20          *zrc20.ZRC20
+	ETHZRC20Addr      ethcommon.Address
+	ETHZRC20          *zrc20.ZRC20
+	BTCZRC20Addr      ethcommon.Address
+	BTCZRC20          *zrc20.ZRC20
+	SOLZRC20Addr      ethcommon.Address
+	SOLZRC20          *zrc20.ZRC20
+	TONZRC20Addr      ethcommon.Address
+	TONZRC20          *zrc20.ZRC20
+	SUIZRC20Addr      ethcommon.Address
+	SUIZRC20          *zrc20.ZRC20
+	SuiTokenZRC20Addr ethcommon.Address
+	SuiTokenZRC20     *zrc20.ZRC20
+
+	// other contracts
 	UniswapV2FactoryAddr ethcommon.Address
 	UniswapV2Factory     *uniswapv2factory.UniswapV2Factory
 	UniswapV2RouterAddr  ethcommon.Address
@@ -243,6 +260,8 @@ func (r *E2ERunner) CopyAddressesFrom(other *E2ERunner) (err error) {
 	r.BTCZRC20Addr = other.BTCZRC20Addr
 	r.SOLZRC20Addr = other.SOLZRC20Addr
 	r.TONZRC20Addr = other.TONZRC20Addr
+	r.SUIZRC20Addr = other.SUIZRC20Addr
+	r.SuiTokenZRC20Addr = other.SuiTokenZRC20Addr
 	r.UniswapV2FactoryAddr = other.UniswapV2FactoryAddr
 	r.UniswapV2RouterAddr = other.UniswapV2RouterAddr
 	r.ConnectorZEVMAddr = other.ConnectorZEVMAddr
@@ -254,6 +273,11 @@ func (r *E2ERunner) CopyAddressesFrom(other *E2ERunner) (err error) {
 	r.ZevmTestDAppAddr = other.ZevmTestDAppAddr
 
 	r.GatewayProgram = other.GatewayProgram
+
+	r.GatewayPackageID = other.GatewayPackageID
+	r.GatewayObjectID = other.GatewayObjectID
+	r.SuiTokenCoinType = other.SuiTokenCoinType
+	r.SuiTokenTreasuryCap = other.SuiTokenTreasuryCap
 
 	// create instances of contracts
 	r.ZetaEth, err = zetaeth.NewZetaEth(r.ZetaEthAddr, r.EVMClient)
@@ -292,6 +316,15 @@ func (r *E2ERunner) CopyAddressesFrom(other *E2ERunner) (err error) {
 	if err != nil {
 		return err
 	}
+	r.SUIZRC20, err = zrc20.NewZRC20(r.SUIZRC20Addr, r.ZEVMClient)
+	if err != nil {
+		return err
+	}
+	r.SuiTokenZRC20, err = zrc20.NewZRC20(r.SuiTokenZRC20Addr, r.ZEVMClient)
+	if err != nil {
+		return err
+	}
+
 	r.UniswapV2Factory, err = uniswapv2factory.NewUniswapV2Factory(r.UniswapV2FactoryAddr, r.ZEVMClient)
 	if err != nil {
 		return err
@@ -365,7 +398,11 @@ func (r *E2ERunner) Unlock() {
 func (r *E2ERunner) PrintContractAddresses() {
 	r.Logger.Print(" --- 📜Solana addresses ---")
 	r.Logger.Print("GatewayProgram: %s", r.GatewayProgram.String())
-	r.Logger.Print("SPL:        %s", r.SPLAddr.String())
+	r.Logger.Print("SPL:            %s", r.SPLAddr.String())
+
+	r.Logger.Print(" --- 📜Sui addresses ---")
+	r.Logger.Print("GatewayPackageID: %s", r.GatewayPackageID)
+	r.Logger.Print("GatewayObjectID:  %s", r.GatewayObjectID)
 
 	// zevm contracts
 	r.Logger.Print(" --- 📜zEVM contracts ---")
@@ -376,6 +413,8 @@ func (r *E2ERunner) PrintContractAddresses() {
 	r.Logger.Print("SOLZRC20:       %s", r.SOLZRC20Addr.Hex())
 	r.Logger.Print("SPLZRC20:       %s", r.SPLZRC20Addr.Hex())
 	r.Logger.Print("TONZRC20:       %s", r.TONZRC20Addr.Hex())
+	r.Logger.Print("SUIZRC20:       %s", r.SUIZRC20Addr.Hex())
+	r.Logger.Print("SuiTokenZRC20:  %s", r.SuiTokenZRC20Addr.Hex())
 	r.Logger.Print("UniswapFactory: %s", r.UniswapV2FactoryAddr.Hex())
 	r.Logger.Print("UniswapRouter:  %s", r.UniswapV2RouterAddr.Hex())
 	r.Logger.Print("ConnectorZEVM:  %s", r.ConnectorZEVMAddr.Hex())